pyproject.toml
[tool.poetry]
name = "dfsync"
version = "0.2.0"
description = ""
authors = ["Mihai Balint <balint.mihai@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
watchdog = "^0.10.2"
gitpython = "^3.1"
kubernetes = "^12.0.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
ipdb = "^0.13.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
